docs: extract styles module
